requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
environs==9.5.0
fastapi==0.109.0
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-google-genai==0.0.6
langsmith==0.0.85
marshmallow==3.20.2
minio==7.2.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.10.0
packaging==23.2
pandas==2.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg==3.1.17
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.0
pydantic_core==2.16.1
pymilvus==2.3.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
